% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.makeTextFun}
\alias{.makeTextFun}
\title{Resolve Summary Function for Focal Operations}
\usage{
.makeTextFun(fun)
}
\arguments{
\item{fun}{Character string or function. Summary function to use.}
}
\value{
A character string or function object to be passed to \code{terra::focal()}.
}
\description{
Converts a summary function input (e.g., "mean", "median", "p25") into either
a valid string recognized by \code{terra::focal()} or a custom function defined below.
}
\keyword{internal}
