% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_progressr.R
\name{set_fastfocal_progress_handler}
\alias{set_fastfocal_progress_handler}
\title{Set the best available progressr handler}
\usage{
set_fastfocal_progress_handler()
}
\description{
Set the best available progressr handler
}
\keyword{internal}
