% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastfocal_weights.R
\name{fastfocal_weights}
\alias{fastfocal_weights}
\title{Generate weight matrix for focal operations using map units}
\usage{
fastfocal_weights(x, d, w = "circle", normalize = TRUE, plot = FALSE)
}
\arguments{
\item{x}{SpatRaster. Used to determine resolution.}

\item{d}{Radius in map units (e.g. meters).}

\item{w}{Type of weight. One of:
"rectangle", "circle", "gaussian", "pareto", "idw", "exponential",
"triangular", "cosine", "logistic", "cauchy", "quartic", "epanechnikov".}

\item{normalize}{Logical. Whether to normalize weights so they sum to 1. Default = TRUE.}

\item{plot}{Logical. If TRUE, plots the resulting kernel. Default = FALSE.}
}
\value{
A numeric matrix of focal weights.
}
\description{
Generate weight matrix for focal operations using map units
}
\examples{
x <- rast(nrows = 100, ncols = 100, xmin = 0, xmax = 3000, ymin = 0, ymax = 3000)
fastfocal_weights(x, d = 300, w = "gaussian", plot = TRUE)
}
